Move layers tab actions into the content area
The add and store buttons were teleported into the library tab strip,
where they overlapped the Canvases tab on narrow widths. Render them in
a right-aligned header row above the layer list instead, matching the
routes tab.
