+++ web/src/views/library/layers/Layers.vue
@@ -10,17 +10,48 @@
 import { computed, onMounted, ref } from 'vue'
 import { storeToRefs } from 'pinia'
 import { useI18n } from 'vue-i18n'
-import { Layers3Icon, StoreIcon } from 'lucide-vue-next'
+import { useRouter } from 'vue-router'
+import {
+  FolderIcon,
+  Layers3Icon,
+  LayersIcon,
+  PlusIcon,
+  StoreIcon,
+} from 'lucide-vue-next'
 import { EmptyState } from '@/components/ui/empty-state'
 import { Button } from '@/components/ui/button'
 import { Spinner } from '@/components/ui/spinner'
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuTrigger,
+} from '@/components/ui/dropdown-menu'
+import { AppRoute } from '@/router'
+import { useAppService } from '@/services/app.service'
 import { useLayersStore } from '@/stores/layers.store'
 import Layers from '@/components/map/Layers.vue'
 import LayerStoreDialog from '@/components/map/layers/LayerStoreDialog.vue'
+import LayerGroupConfiguration from '@/components/map/layers/LayerGroupConfiguration.vue'
 
 const { t } = useI18n()
+const router = useRouter()
+const appService = useAppService()
 const layersStore = useLayersStore()
 const { mainReorderableItems } = storeToRefs(layersStore)
+
+// The layer editor is a sheet view, not a dialog: it renders the draft on
+// the live map while you edit, which a modal over the map cannot do.
+function newLayer() {
+  router.push({ name: AppRoute.LAYER_EDITOR_NEW })
+}
+
+function newGroup() {
+  appService.componentDialog({
+    component: LayerGroupConfiguration,
+    continueText: t('general.save'),
+  })
+}
 
 const loading = ref(mainReorderableItems.value.length === 0)
 const storeOpen = ref(false)
@@ -52,8 +83,40 @@ const isEmpty = computed(
         {{ t('layers.store.title') }}
       </Button>
     </EmptyState>
-    <LayerStoreDialog v-model:open="storeOpen" />
   </div>
 
-  <Layers v-else />
+  <div v-else class="h-full flex flex-col gap-2">
+    <div class="flex items-center justify-end gap-2">
+      <Button
+        variant="outline"
+        size="icon"
+        class="h-10 w-10"
+        :aria-label="t('layers.store.title')"
+        :title="t('layers.store.title')"
+        @click="storeOpen = true"
+      >
+        <StoreIcon class="h-4 w-4" />
+      </Button>
+      <DropdownMenu>
+        <DropdownMenuTrigger as-child>
+          <Button variant="outline" size="icon" class="h-10 w-10">
+            <PlusIcon class="h-4 w-4" />
+          </Button>
+        </DropdownMenuTrigger>
+        <DropdownMenuContent align="end">
+          <DropdownMenuItem @click="newLayer">
+            <LayersIcon class="size-4" />
+            {{ t('layers.actions.newLayer') }}
+          </DropdownMenuItem>
+          <DropdownMenuItem @click="newGroup">
+            <FolderIcon class="size-4" />
+            {{ t('layers.actions.newGroup') }}
+          </DropdownMenuItem>
+        </DropdownMenuContent>
+      </DropdownMenu>
+    </div>
+    <Layers class="flex-1 min-h-0" />
+  </div>
+
+  <LayerStoreDialog v-model:open="storeOpen" />
 </template>
